Bugfix + Change multiple vocables feature
Fixed: When inserting a special character into a text field, the caret
would jump back to position 0, instead of a more appropriate position,
so that the edited part of the text of the text field is always visible.
Added feature: When highlighting multiple vocables in the vocable table,
you can now change them in one dialogue,instead of clicking the change
vocable button once per vocable.